requirements.txt
fastapi
uvicorn
supabase
python-dotenv
google-generativeai
python-multipart
numpy
# face_recognition
opencv-python
